BhawanUppal/

+++ WJets/8TeV/LeptoQuarks/NTupleMaker/src/RootTupleMakerV2_Event.cc
@@ -14,14 +14,14 @@ RootTupleMakerV2_Event::RootTupleMakerV2_Event(const edm::ParameterSet& iConfig)
   produces <unsigned int> ( "bunch"  );
   produces <unsigned int> ( "ls"     );
   produces <unsigned int> ( "orbit"  );
-  produces <double>       ( "time"   );
+  produces <float>       ( "time"   );
   produces <bool>         ( "isData" );
   //produces <double>     ( "rhoIso" );//not used in 2012 
-  produces <double>       ( "rhoJets"     );
-  produces <double>       ( "rhoForHEEP"  );
-  produces <double>       ( "rhoJetsCCPU" );
-  produces <double>       ( "rhoJetsCN"   );
-  produces <double>       ( "rhoJetsCNT"  );
+  produces <float>       ( "rhoJets"     );
+  produces <float>       ( "rhoForHEEP"  );
+  produces <float>       ( "rhoJetsCCPU" );
+  produces <float>       ( "rhoJetsCN"   );
+  produces <float>       ( "rhoJetsCNT"  );
 }
 
 void RootTupleMakerV2_Event::
@@ -31,13 +31,13 @@ produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
   std::auto_ptr<unsigned int >  event ( new unsigned int(iEvent.id().event()      ) );
   std::auto_ptr<unsigned int >  ls    ( new unsigned int(iEvent.luminosityBlock() ) );
 
-  double sec  = iEvent.time().value() >> 32 ;
-  double usec = 0xFFFFFFFF & iEvent.time().value();
-  double conv = 1e6;
+  float sec  = iEvent.time().value() >> 32 ;
+  float usec = 0xFFFFFFFF & iEvent.time().value();
+  float conv = 1e6;
 
   std::auto_ptr<unsigned int >  bunch ( new unsigned int(iEvent.bunchCrossing()   ) );
   std::auto_ptr<unsigned int >  orbit ( new unsigned int(iEvent.orbitNumber()     ) );
-  std::auto_ptr<double >        time  ( new double(sec+usec/conv));
+  std::auto_ptr<float >        time  ( new float(sec+usec/conv));
 
   std::auto_ptr<bool >          isdata  ( new bool(iEvent.isRealData()));
 
@@ -47,23 +47,23 @@ produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
 
   edm::Handle<double> rhoHJets;
   iEvent.getByLabel(fastJetForJECInputTag,rhoHJets);
-  std::auto_ptr<double >        rhoJets     ( new double( *rhoHJets.product() )     );
+  std::auto_ptr<float >        rhoJets     ( new float( *rhoHJets.product() )     );
 
   edm::Handle<double> rhoHForHEEP;
   iEvent.getByLabel(fastJetForHEEPInputTag, rhoHForHEEP);
-  std::auto_ptr<double >        rhoForHEEP  ( new double( *rhoHForHEEP.product() )     );
+  std::auto_ptr<float >        rhoForHEEP  ( new float( *rhoHForHEEP.product() )     );
 
   edm::Handle<double> rhoHJetsCCPU;
   iEvent.getByLabel(fastJetForJECCCPUInputTag,rhoHJetsCCPU);
-  std::auto_ptr<double >        rhoJetsCCPU ( new double( *rhoHJetsCCPU.product() ) );
+  std::auto_ptr<float >        rhoJetsCCPU ( new float( *rhoHJetsCCPU.product() ) );
 
   edm::Handle<double> rhoHJetsCN;
   iEvent.getByLabel(fastJetForJECCNInputTag,rhoHJetsCN);
-  std::auto_ptr<double >        rhoJetsCN   ( new double( *rhoHJetsCN.product() )   );
+  std::auto_ptr<float >        rhoJetsCN   ( new float( *rhoHJetsCN.product() )   );
 
   edm::Handle<double> rhoHJetsCNT;
   iEvent.getByLabel(fastJetForJECCNTInputTag,rhoHJetsCNT);
-  std::auto_ptr<double >        rhoJetsCNT  ( new double( *rhoHJetsCNT.product() )  );
+  std::auto_ptr<float >        rhoJetsCNT  ( new float( *rhoHJetsCNT.product() )  );
